pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "discord-py>=2.5.0",
    "requests>=2.32.3",
    "trafilatura>=2.0.0",
]
